+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - Changed way of creating User - from now Authority will be fetched from DB while creating User
+- Removed Parent IDs in patch Student request - flow does not require student side update, Parent cannot be created without children and updates both sides of relation
 
 ## [1.3.2] - 2026-01-03
 ### Fixed